Use an arena to store objects on the heap
This changes the ownership strategy being used for acceleration
structures, since it doesn't make much sense to have them own the
objects in the scene. The ownership is now tied to the `Renderer` struct
which holds the state for the entire renderer, and the acceleration
structure simply holds a smart pointer to it.

This also exposes a convenient `Arena` type alias.